services/style-transfer/requirements.txt
Flask==1.1.2
flask-cors==3.0.10
tensorflow
tensorflow-hub
Pillow==8.2.0
